pyproject.toml
[project]
name = "askari-patrol-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.121.3",
    "httpx>=0.28.1",
    "mcp[cli]>=1.21.1",
    "prompt-toolkit>=3.0.52",
    "pydantic>=2.12.4",
    "pydantic-ai-slim[anthropic,google,groq,logfire]>=1.22.0",
    "pyjwt>=2.10.1",
    "pytest-asyncio>=1.3.0",
    "respx>=0.22.0",
    "rich>=14.2.0",
    "twilio>=9.8.7",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = ["pytest>=9.0.1", "pytest-cov>=7.0.0", "ruff>=0.14.6"]

[tool.pytest.ini_options]
addopts = "--maxfail=1"
pythonpath = ["src"]


[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501", # line length (handled by formatter)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
